fixed size_t string size
git-svn-id: svn+ssh://cdcvs.fnal.gov/cvs/projects/damic/stable/writeDamicHeader@30 678ac9d6-6348-4232-9a94-0a0f05251285

+++ writeDamicHeader.cc
@@ -44,7 +44,7 @@ bool readStatsFile(const string &statsFile, map<string,double> &mCards){
     in.getline(line,kMaxLine);
     
     string lineS(line);
-    unsigned int found = lineS.find_first_not_of(" \t\v\r\n");
+    size_t found = lineS.find_first_not_of(" \t\v\r\n");
     if(found == string::npos) continue;
     else if( line[found] == '#') continue;
